+++ example/src/auth0-configuration.d.ts
@@ -1,0 +1,9 @@
+interface Auth0Configuration {
+  domain: string;
+  clientId: string;
+  domainSecondScreen: string;
+  clientIdSecondScreen: string;
+}
+
+declare const config: Auth0Configuration;
+export default config;