+++ src/utils/whitelist.ts
@@ -8,7 +8,7 @@ export default class ParameterError extends BaseError {
   constructor(expected: unknown[], actual: string, missing: unknown[]) {
     super(
       'Missing required parameters',
-      `Missing required parameters: ${JSON.stringify(missing, null, 2)}`,
+      `Missing required parameters: ${JSON.stringify(missing, null, 2)}`
     );
     this.expected = expected;
     this.actual = actual;
@@ -17,11 +17,11 @@ export default class ParameterError extends BaseError {
 }
 
 export function apply(rules: any, values: any) {
-  const {whitelist = true, parameters, aliases = {}} = rules;
-  let mapped: {[key: string]: string} = {};
+  const { whitelist = true, parameters, aliases = {} } = rules;
+  let mapped: { [key: string]: string } = {};
   let requiredKeys = Object.keys(parameters)
-    .filter(key => parameters[key].required)
-    .map(key => parameters[key].toName || key);
+    .filter((key) => parameters[key].required)
+    .map((key) => parameters[key].toName || key);
   for (let key of Object.keys(values)) {
     let value = values[key];
     let parameterKey: string = aliases[key] || key;
@@ -32,7 +32,7 @@ export function apply(rules: any, values: any) {
       mapped[key] = value;
     }
   }
-  let missing = requiredKeys.filter(key => !mapped[key]);
+  let missing = requiredKeys.filter((key) => !mapped[key]);
   if (missing.length > 0) {
     throw new ParameterError(requiredKeys, values, missing);
   }